# Move `bin/` and `tests/` out of ipa-core