--- a/docker/ci/helper.Dockerfile
+++ b/docker/ci/helper.Dockerfile
@@ -3,8 +3,7 @@ FROM rust:latest as builder
 
 COPY . /ipa/
 RUN cd /ipa && \
-  cargo build --bin helper --release --no-default-features \
-        --features "web-app real-world-infra compact-gate"
+  cargo build --bin helper --release --features "helper ipa-core/compact-gate"
 
 # Copy them to the final image
 FROM debian:bullseye-slim