--- a/ipa-test/tests/helper_networks.rs
+++ b/ipa-test/tests/helper_networks.rs
@@ -9,25 +9,21 @@ use common::{
 use ipa_core::{cli::CliPaths, helpers::HelperIdentity, test_fixture::ipa::IpaSecurityModel};
 
 #[test]
-#[cfg(all(test, web_test))]
 fn http_network() {
     test_network(false);
 }
 
 #[test]
-#[cfg(all(test, web_test))]
 fn https_network() {
     test_network(true);
 }
 
 #[test]
-#[cfg(all(test, web_test))]
 fn http_semi_honest_ipa() {
     test_ipa(IpaSecurityModel::SemiHonest, false);
 }
 
 #[test]
-#[cfg(all(test, web_test))]
 fn https_semi_honest_ipa() {
     test_ipa(IpaSecurityModel::SemiHonest, true);
 }
@@ -37,7 +33,6 @@ fn https_semi_honest_ipa() {
 ///
 /// [`network`]: crate::test_network
 #[test]
-#[cfg(all(test, web_test))]
 fn keygen_confgen() {
     let dir = TempDir::new_delete_on_drop();
     let path = dir.path();