--- a/pre-commit
+++ b/pre-commit
@@ -79,31 +79,31 @@ check() {
 }
 
 check "Benchmark compilation" \
-    cargo build --benches --no-default-features --features "enable-benches descriptive-gate"
+    cargo build --benches --features "enable-benches descriptive-gate"
 
 check "Clippy checks" \
-    cargo clippy --tests -- -D warnings
+    cargo clippy --tests --features basic -- -D warnings
 
 check "Clippy concurrency checks" \
-    cargo clippy --tests --features shuttle -- -D warnings
+    cargo clippy --tests --features "basic shuttle" -- -D warnings
 
 check "Tests" \
-    cargo test
+    cargo test --features basic
 
 check "Web tests" \
-    cargo test -p ipa-core --no-default-features --features "cli web-app real-world-infra test-fixture descriptive-gate"
+    cargo test --features "helper ipa-core/descriptive-gate integration-test"
 
 check "Concurrency tests" \
-    cargo test -p ipa-core --release --features shuttle
+    cargo test --release --features "basic shuttle"
 
 check "IPA benchmark" \
-    cargo bench --bench oneshot_ipa --no-default-features --features="enable-benches descriptive-gate" -- -n 62
+    cargo bench --bench oneshot_ipa --features="enable-benches descriptive-gate" -- -n 62
 
 check "IPA OPRF benchmark" \
-    cargo bench --bench oneshot_ipa --no-default-features --features="enable-benches descriptive-gate" -- -n 62 --oprf -c 16
+    cargo bench --bench oneshot_ipa --features="enable-benches descriptive-gate" -- -n 62 --oprf -c 16
 
 check "Arithmetic circuit benchmark" \
-    cargo bench --bench oneshot_arithmetic --no-default-features --features "enable-benches descriptive-gate"
+    cargo bench --bench oneshot_arithmetic --features "enable-benches descriptive-gate"
 
 check "Sort benchmark" \
-    cargo bench --bench oneshot_sort --no-default-features --features="enable-benches descriptive-gate"
+    cargo bench --bench oneshot_sort --features="enable-benches descriptive-gate"